Fix second-level compact unwind table corruption
The issue would occur when the page is ended because of no
more space for page-specific encodings. In that case, the
loop would terminate (!canDo case 3) with the entryCount off
by one. Thus, when filling the encoding table afterwards, we
would read 0 for the index of that encoding and write before
the start of encodingsArray corrupting the entriesArray.